Move unit tests to a separate project
- Moved unit tests from `Program.cs` to a new test project `TestProject` for better organization and modularity.
- Updated the solution `sharp-grad.sln` to include the new test project and its build configurations.
