+++ sharp-grad.sln
@@ -1,4 +1,3 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34728.123
@@ -6,6 +5,8 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpGrad", "SharpGrad\SharpGrad.csproj", "{1909F459-32FA-4466-9C15-46CEF50C3904}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Example", "Example\Example.csproj", "{4BE524C8-3F15-4E9D-9566-D27C588A52C8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject", "TestProject\TestProject.csproj", "{8B5AC941-E34E-4BAF-A253-461830E6FA44}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +22,10 @@ Global
 		{4BE524C8-3F15-4E9D-9566-D27C588A52C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4BE524C8-3F15-4E9D-9566-D27C588A52C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4BE524C8-3F15-4E9D-9566-D27C588A52C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B5AC941-E34E-4BAF-A253-461830E6FA44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B5AC941-E34E-4BAF-A253-461830E6FA44}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B5AC941-E34E-4BAF-A253-461830E6FA44}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B5AC941-E34E-4BAF-A253-461830E6FA44}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
